+++ src/xin_fs/xin_note.c
@@ -14,12 +14,12 @@ void note_input(void)
 
     uint8_t selected_character;
 
-    if(&Screen.cursor[Screen.y][Screen.x] < VGA_TEXT_MEMORY)
+    if((uint32_t)&Screen.cursor[Screen.y][Screen.x] < VGA_TEXT_MEMORY)
     {
         Screen.y++;
     }
 
-    if(&Screen.cursor[Screen.y][Screen.x] > VGA_TEXT_MEMORY + SCREEN_RESOLUTION)
+    if((uint32_t)&Screen.cursor[Screen.y][Screen.x] > VGA_TEXT_MEMORY + SCREEN_RESOLUTION)
     {
         Screen.y--;
     }
