+++ src/xin_fs/xin_paint.c
@@ -16,13 +16,13 @@ void paint_input(void)
 
     uint16_t selected_cell;
 
-    if(&Screen.cursor[Screen.y][Screen.x] < VGA_TEXT_MEMORY)
+    if((uint32_t)&Screen.cursor[Screen.y][Screen.x] < VGA_TEXT_MEMORY)
     {
         Screen.y = 0;
         Screen.x = 0;
     }
 
-    if(&Screen.cursor[Screen.y][Screen.x] > VGA_TEXT_MEMORY + VGA_SCREEN_RESOLUTION)
+    if((uint32_t)&Screen.cursor[Screen.y][Screen.x] > VGA_TEXT_MEMORY + VGA_SCREEN_RESOLUTION)
     {
         Screen.y = 0;
         Screen.x = 0;
